Remove extra lifecyle functions to fix pacman keyboard movement
